Add files via upload
This project consists in a web application titled “Dreamcatcher”. Here the user can introduce a description of a dream they had and also rate the duration, stress level and energy of their sleep out of 5, information stored in a database. With a combo-box each of this three metrics can be selected and then the user will be redirected to a new page that show the chart for the attribute selected. I chose Python to develop this application, using the Django framework and HTML for the front-end component.